+++ glfw/init.c
@@ -57,7 +57,10 @@ static _GLFWinitconfig _glfwInitHints = {
     .ns = {
         .menubar = true,  // macOS menu bar
         .chdir = true   // macOS bundle chdir
-    }
+    },
+    .wl = {
+        .ime = true,  // Wayland IME support
+    },
 };
 
 // Terminate the library
@@ -295,6 +298,9 @@ GLFWAPI void glfwInitHint(int hint, int value)
             return;
         case GLFW_COCOA_MENUBAR:
             _glfwInitHints.ns.menubar = value;
+            return;
+        case GLFW_WAYLAND_IME:
+            _glfwInitHints.wl.ime = value;
             return;
     }