+++ kitty/options/to-c-generated.h
@@ -1123,6 +1123,19 @@ convert_from_opts_macos_colorspace(PyObject *py_opts, Options *opts) {
     Py_DECREF(ret);
 }
 
+static void
+convert_from_python_wayland_enable_ime(PyObject *val, Options *opts) {
+    opts->wayland_enable_ime = PyObject_IsTrue(val);
+}
+
+static void
+convert_from_opts_wayland_enable_ime(PyObject *py_opts, Options *opts) {
+    PyObject *ret = PyObject_GetAttrString(py_opts, "wayland_enable_ime");
+    if (ret == NULL) return;
+    convert_from_python_wayland_enable_ime(ret, opts);
+    Py_DECREF(ret);
+}
+
 static bool
 convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     convert_from_opts_font_size(py_opts, opts);
@@ -1296,6 +1309,8 @@ convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     convert_from_opts_macos_menubar_title_max_length(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_macos_colorspace(py_opts, opts);
+    if (PyErr_Occurred()) return false;
+    convert_from_opts_wayland_enable_ime(py_opts, opts);
     if (PyErr_Occurred()) return false;
     return true;
 }